requirements.txt

arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
cssselect2==0.7.0
django-environ
Django==5.0.1
django-admin-interface==0.28.3
django-colorfield==0.11.0
django-environ==0.11.2
freetype-py==2.3.0
html5lib==1.1
idna==3.6
lxml==5.1.0
oscrypto==1.3.0
pillow==10.2.0
psycopg==3.1.17
psycopg2-binary==2.9.9
pycairo==1.25.1
pycparser==2.21
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
pypdf==3.17.4
pypng==0.20220715.0
python-bidi==0.4.2
python-slugify==8.0.1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.9
requests==2.31.0
rlPyCairo==0.3.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
text-unidecode==1.3
tinycss2==1.2.1
typing_extensions==4.9.0
tzlocal==5.2
uritools==4.0.2
urllib3==2.1.0
webencodings==0.5.1
xhtml2pdf==0.2.13
